+++ subprojects/zap-clientapi/src/main/java/org/zaproxy/clientapi/gen/deprecated/SpiderDeprecated.java
@@ -548,4 +548,28 @@ public class SpiderDeprecated {
         map.put("Integer", Integer.toString(i));
         return api.callApi("spider", "action", "setOptionThreadCount", map);
     }
+
+    /** @deprecated Option no longer in effective use. */
+    @Deprecated
+    public ApiResponse optionScope() throws ClientApiException {
+        return api.callApi("spider", "view", "optionScope", null);
+    }
+
+    /** @deprecated Option no longer in effective use. */
+    @Deprecated
+    public ApiResponse optionScopeText() throws ClientApiException {
+        return api.callApi("spider", "view", "optionScopeText", null);
+    }
+
+    /**
+     * Use actions [add|modify|remove]DomainAlwaysInScope instead.
+     *
+     * @deprecated Option no longer in effective use.
+     */
+    @Deprecated
+    public ApiResponse setOptionScopeString(String string) throws ClientApiException {
+        Map<String, String> map = new HashMap<>();
+        map.put("String", string);
+        return api.callApi("spider", "action", "setOptionScopeString", map);
+    }
 }